Moved Commonly used Cards to their own Folder
