+++ frontend/components/profile/UserProfile/UserPodcasts.tsx
@@ -23,7 +23,7 @@ import NextLink from "next/link";
 import { Podcast } from "../../../utilities/Interfaces";
 import AuthHelper from "../../../helpers/AuthHelper";
 import PodcastHelper from "../../../helpers/PodcastHelper";
-import PodcastCard from "../../explore/PodcastCard";
+import PodcastCard from "../../Cards/PodcastCard";
 
 export default function UserPodcasts({ userId }) {
   // podcasts data
