+++ frontend/components/shared/Navbar.tsx
@@ -39,8 +39,8 @@ export default function Navbar() {
   const { data: session, status } = useSession();
   const { colorMode, toggleColorMode } = useColorMode();
   const isMobile = useBreakpointValue({ base: true, md: false });
+
   const [searchValue, setSearchValue] = useState("");
-  const handleSearchChange = (event) => setSearchValue(event.target.value);
 
   const handleSearchSubmit = () => {
     const searchlink = "/Explore/Search?searchTerm=" + searchValue;
@@ -119,6 +119,10 @@ export default function Navbar() {
     setIsUserLoggedIn(false);
     setIsUserSet(false);
     window.location.href = indexPage;
+  };
+
+  const handleSearchChange = (event) => {
+    setSearchValue(event.target.value);
   };
 
   /**
