+++ frontend/pages/Create.tsx
@@ -63,7 +63,7 @@ const CreateEpisode = () => {
       // If logged in, set user, otherwise redirect to login page
       if (res.status == 200) {
         setUser(res.userMenuInfo);
-        PodcastHelper.podcastMyPodcastsGet().then((res2) => {
+        PodcastHelper.podcastMyPodcastsGet(0, 12).then((res2) => {
           // If logged in, set user, otherwise redirect to login page
           if (res2.status == 200) {
             setPodcasts(res2.myPodcasts);
@@ -157,301 +157,305 @@ const CreateEpisode = () => {
     router.push("/NewPodcast");
   };
 
-  return <>
-    <Navbar />
-    <Center>
-      <VStack mt={"1em"}>
-        <Heading fontWeight={"normal"} fontFamily={"Avenir Next"}>
-          Upload Episode
-        </Heading>
-        <Text mb={"1em"}>Choose a Podcast</Text>
-      </VStack>
-    </Center>
+  return (
+    <>
+      <Navbar />
+      <Center>
+        <VStack mt={"1em"}>
+          <Heading fontWeight={"normal"} fontFamily={"Avenir Next"}>
+            Upload Episode
+          </Heading>
+          <Text mb={"1em"}>Choose a Podcast</Text>
+        </VStack>
+      </Center>
 
-    <Box display="flex" width="100%" justifyContent="center">
-      <Flex
-        direction="row"
-        wrap="wrap"
-        justifyContent="center"
-        alignItems="center"
-      >
-        {podcasts.map((podcast) => (
+      <Box display="flex" width="100%" justifyContent="center">
+        <Flex
+          direction="row"
+          wrap="wrap"
+          justifyContent="center"
+          alignItems="center"
+        >
+          {podcasts.map((podcast) => (
+            <Flex
+              key={podcast.id}
+              direction="column"
+              alignItems="center"
+              bg={
+                selectedPodcast && selectedPodcast.id === podcast.id
+                  ? colorMode === "light"
+                    ? "#00000010"
+                    : "#FFFFFF10"
+                  : "transparent"
+              }
+              borderRadius="1em"
+              cursor="pointer"
+              outline={
+                selectedPodcast && selectedPodcast.id === podcast.id
+                  ? "2px solid #3182CE"
+                  : "none"
+              }
+              onClick={() => handlePodcastSelect(podcast)}
+              p={2}
+              m={2}
+            >
+              <Image
+                src={podcast.coverArtUrl}
+                alt={podcast.name}
+                boxSize="100px"
+                borderRadius="2em"
+                objectFit="cover"
+                boxShadow="lg"
+                outline="2px solid #FFFFFF80"
+                data-cy={`podcast-image-${podcast.name
+                  .replace(/\s+/g, "-")
+                  .toLowerCase()}`} // Adding a data-cy attribute to the Image component
+              />
+              <Text mt={2}>
+                {" "}
+                {podcast.name.length > 18
+                  ? `${podcast.name.substring(0, 18)}...`
+                  : podcast.name}
+              </Text>
+            </Flex>
+          ))}
+
+          {/* Create a Podcast Option */}
           <Flex
-            key={podcast.id}
             direction="column"
             alignItems="center"
-            bg={
-              selectedPodcast && selectedPodcast.id === podcast.id
-                ? colorMode === "light"
-                  ? "#00000010"
-                  : "#FFFFFF10"
-                : "transparent"
-            }
             borderRadius="1em"
             cursor="pointer"
-            outline={
-              selectedPodcast && selectedPodcast.id === podcast.id
-                ? "2px solid #3182CE"
-                : "none"
-            }
-            onClick={() => handlePodcastSelect(podcast)}
+            outline="none"
+            onClick={navigateToCreatePodcast}
             p={2}
             m={2}
+            bg="transparent"
           >
-            <Image
-              src={podcast.coverArtUrl}
-              alt={podcast.name}
+            <Box
               boxSize="100px"
               borderRadius="2em"
-              objectFit="cover"
-              boxShadow="lg"
-              outline="2px solid #FFFFFF80"
-              data-cy={`podcast-image-${podcast.name.replace(/\s+/g, '-').toLowerCase()}`} // Adding a data-cy attribute to the Image component
-            />
-            <Text mt={2}>
-              {" "}
-              {podcast.name.length > 18
-                ? `${podcast.name.substring(0, 18)}...`
-                : podcast.name}
-            </Text>
-          </Flex>
-        ))}
-
-        {/* Create a Podcast Option */}
-        <Flex
-          direction="column"
-          alignItems="center"
-          borderRadius="1em"
-          cursor="pointer"
-          outline="none"
-          onClick={navigateToCreatePodcast}
-          p={2}
-          m={2}
-          bg="transparent"
-        >
-          <Box
-            boxSize="100px"
-            borderRadius="2em"
-            border="2px dashed gray"
-            display="flex"
-            alignItems="center"
-            justifyContent="center"
-          >
-            <AddIcon w={8} h={8} />
-          </Box>
-          <Text mt={2}>Create a Podcast</Text>
-        </Flex>
-      </Flex>
-    </Box>
-
-    {/* Form Container */}
-    <Box display="flex" justifyContent="center">
-      <VStack spacing={5} align="center" p={5}>
-        {/* Form Container */}
-        {/* Displaying Selected Podcast Title */}
-        {selectedPodcast && (
-          <Text
-            fontSize="xl"
-            fontWeight="normal"
-            bg={
-              colorMode === "light"
-                ? "rgba(0, 0, 0, 0.1)"
-                : "rgba(255, 255, 255, 0.1)"
-            } // Slight transparency
-            pl={5} // Padding for better visual spacing
-            pr={5} // Padding for better visual spacing
-            pt={2} // Padding for better visual spacing
-            pb={2} // Padding for better visual spacing
-            mb={5} // Margin for better visual spacing
-            borderRadius="5em" // Rounded corners
-            outline={
-              colorMode === "light"
-                ? "1px solid #000000"
-                : "1px solid #FFFFFF"
-            } // Black or white border
-            fontFamily={"Avenir Next"}
-          >
-            {`${selectedPodcast.name}`}
-          </Text>
-        )}
-        <form onSubmit={handleAdd}>
-          <div
-            style={{
-              position: "relative",
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "center",
-            }}
-          >
-            <img
-              src={
-                coverImage ||
-                "https://img.icons8.com/?size=512&id=492ILERveW8G&format=png"
-              }
-              alt="Cover Photo"
-              style={{
-                width: "150px",
-                height: "150px",
-                borderRadius: "50%",
-                padding: "15px",
-                position: "relative",
-              }}
-            />
-            <label
-              htmlFor="Cover Photo"
-              style={{
-                position: "absolute",
-                cursor: "pointer",
-                bottom: "15px",
-                right: "5px",
-              }}
-            >
-              <IconButton
-                aria-label="Upload Cover Photo"
-                icon={
-                  <img
-                    src="https://img.icons8.com/?size=512&id=hwKgsZN5Is2H&format=png"
-                    alt="Upload Icon"
-                    width="25px"
-                    height="25px"
-                  />
-                }
-                size="sm"
-                variant="outline"
-                borderRadius="full"
-                border="1px solid grey"
-                padding={3}
-                style={{
-                  backdropFilter: "blur(5px)",
-                  backgroundColor: "rgba(0, 0, 0, 0.4)",
-                }}
-                zIndex={-999}
-              />
-              <input
-                type="file"
-                id="Cover Photo"
-                accept="image/*"
-                onChange={(e) => handleCoverImageUpload(e)}
-                style={{
-                  display: "none",
-                }}
-              />
-            </label>
-          </div>
-          {addError && <Text color="red.500">{addError}</Text>}
-          <VStack spacing={5} align="center" p={5}>
-            {/* Episode Name Input */}
-            <FormControl position="relative">
-              <Input
-                value={episodeName}
-                onChange={handleEpisodeNameChange}
-                placeholder="Enter episode name..."
-                rounded="lg"
-                pr="50px"
-              />
-              <Text
-                position="absolute"
-                right="8px"
-                bottom="8px"
-                fontSize="sm"
-                color="gray.500"
-              >
-                {episodeNameCharacterCount}/25
-              </Text>
-            </FormControl>
-
-            {/* Description Textarea */}
-            <FormControl position="relative">
-              <Textarea
-                value={description}
-                onChange={handleDescriptionChange}
-                placeholder="Enter episode description..."
-              />
-              <Text
-                position="absolute"
-                right="8px"
-                bottom="8px"
-                fontSize="sm"
-                color="gray.500"
-              >
-                {descriptionCharacterCount}/250
-              </Text>
-            </FormControl>
-
-            {/* Genre Selection */}
-            <FormControl
+              border="2px dashed gray"
               display="flex"
               alignItems="center"
               justifyContent="center"
             >
-              <Switch
-                id="explicitToggle"
-                colorScheme="purple"
-                isChecked={isExplicit}
-                onChange={() => setIsExplicit(!isExplicit)}
-                opacity={0.9} // Setting the opacity to 0.7 to make it slightly faded
-              >
-                Explicit
-              </Switch>
-            </FormControl>
-
-            {/* File Upload */}
-            <Box
-              {...getRootProps()}
-              border="2px dashed gray"
-              borderRadius="md"
-              p={4}
-              textAlign="center"
-              width="300px"
-            >
-              <input {...getInputProps()} />
-              {file ? (
-                <p>{file.name}</p>
-              ) : (
-                <p>Drag & drop a podcast file here, or click to select one</p>
-              )}
+              <AddIcon w={8} h={8} />
             </Box>
+            <Text mt={2}>Create a Podcast</Text>
+          </Flex>
+        </Flex>
+      </Box>
 
-            {/* Upload Button */}
-            <Button
-              id="createBtn"
-              type="submit"
-              fontSize="md"
-              borderRadius={"full"}
-              minWidth={"200px"}
-              color={"white"}
-              marginTop={"15px"}
-              marginBottom={"10px"}
-              padding={"20px"}
-              // semi transparent white outline
-              outline={"1px solid rgba(255, 255, 255, 0.6)"}
+      {/* Form Container */}
+      <Box display="flex" justifyContent="center">
+        <VStack spacing={5} align="center" p={5}>
+          {/* Form Container */}
+          {/* Displaying Selected Podcast Title */}
+          {selectedPodcast && (
+            <Text
+              fontSize="xl"
+              fontWeight="normal"
+              bg={
+                colorMode === "light"
+                  ? "rgba(0, 0, 0, 0.1)"
+                  : "rgba(255, 255, 255, 0.1)"
+              } // Slight transparency
+              pl={5} // Padding for better visual spacing
+              pr={5} // Padding for better visual spacing
+              pt={2} // Padding for better visual spacing
+              pb={2} // Padding for better visual spacing
+              mb={5} // Margin for better visual spacing
+              borderRadius="5em" // Rounded corners
+              outline={
+                colorMode === "light"
+                  ? "1px solid #000000"
+                  : "1px solid #FFFFFF"
+              } // Black or white border
+              fontFamily={"Avenir Next"}
+            >
+              {`${selectedPodcast.name}`}
+            </Text>
+          )}
+          <form onSubmit={handleAdd}>
+            <div
               style={{
-                background:
-                  "linear-gradient(45deg, #007BFF, #3F60D9, #5E43BA, #7C26A5, #9A0A90)",
-                backgroundSize: "300% 300%",
-                animation: "Gradient 10s infinite linear",
+                position: "relative",
+                display: "flex",
+                flexDirection: "column",
+                alignItems: "center",
               }}
             >
-              Upload
-              <style jsx>{`
-                @keyframes Gradient {
-                  0% {
-                    background-position: 100% 0%;
-                  }
-                  50% {
-                    background-position: 0% 100%;
-                  }
-                  100% {
-                    background-position: 100% 0%;
-                  }
+              <img
+                src={
+                  coverImage ||
+                  "https://img.icons8.com/?size=512&id=492ILERveW8G&format=png"
                 }
-              `}</style>
-            </Button>
-          </VStack>
-        </form>
-      </VStack>
-    </Box>
-  </>;
+                alt="Cover Photo"
+                style={{
+                  width: "150px",
+                  height: "150px",
+                  borderRadius: "50%",
+                  padding: "15px",
+                  position: "relative",
+                }}
+              />
+              <label
+                htmlFor="Cover Photo"
+                style={{
+                  position: "absolute",
+                  cursor: "pointer",
+                  bottom: "15px",
+                  right: "5px",
+                }}
+              >
+                <IconButton
+                  aria-label="Upload Cover Photo"
+                  icon={
+                    <img
+                      src="https://img.icons8.com/?size=512&id=hwKgsZN5Is2H&format=png"
+                      alt="Upload Icon"
+                      width="25px"
+                      height="25px"
+                    />
+                  }
+                  size="sm"
+                  variant="outline"
+                  borderRadius="full"
+                  border="1px solid grey"
+                  padding={3}
+                  style={{
+                    backdropFilter: "blur(5px)",
+                    backgroundColor: "rgba(0, 0, 0, 0.4)",
+                  }}
+                  zIndex={-999}
+                />
+                <input
+                  type="file"
+                  id="Cover Photo"
+                  accept="image/*"
+                  onChange={(e) => handleCoverImageUpload(e)}
+                  style={{
+                    display: "none",
+                  }}
+                />
+              </label>
+            </div>
+            {addError && <Text color="red.500">{addError}</Text>}
+            <VStack spacing={5} align="center" p={5}>
+              {/* Episode Name Input */}
+              <FormControl position="relative">
+                <Input
+                  value={episodeName}
+                  onChange={handleEpisodeNameChange}
+                  placeholder="Enter episode name..."
+                  rounded="lg"
+                  pr="50px"
+                />
+                <Text
+                  position="absolute"
+                  right="8px"
+                  bottom="8px"
+                  fontSize="sm"
+                  color="gray.500"
+                >
+                  {episodeNameCharacterCount}/25
+                </Text>
+              </FormControl>
+
+              {/* Description Textarea */}
+              <FormControl position="relative">
+                <Textarea
+                  value={description}
+                  onChange={handleDescriptionChange}
+                  placeholder="Enter episode description..."
+                />
+                <Text
+                  position="absolute"
+                  right="8px"
+                  bottom="8px"
+                  fontSize="sm"
+                  color="gray.500"
+                >
+                  {descriptionCharacterCount}/250
+                </Text>
+              </FormControl>
+
+              {/* Genre Selection */}
+              <FormControl
+                display="flex"
+                alignItems="center"
+                justifyContent="center"
+              >
+                <Switch
+                  id="explicitToggle"
+                  colorScheme="purple"
+                  isChecked={isExplicit}
+                  onChange={() => setIsExplicit(!isExplicit)}
+                  opacity={0.9} // Setting the opacity to 0.7 to make it slightly faded
+                >
+                  Explicit
+                </Switch>
+              </FormControl>
+
+              {/* File Upload */}
+              <Box
+                {...getRootProps()}
+                border="2px dashed gray"
+                borderRadius="md"
+                p={4}
+                textAlign="center"
+                width="300px"
+              >
+                <input {...getInputProps()} />
+                {file ? (
+                  <p>{file.name}</p>
+                ) : (
+                  <p>Drag & drop a podcast file here, or click to select one</p>
+                )}
+              </Box>
+
+              {/* Upload Button */}
+              <Button
+                id="createBtn"
+                type="submit"
+                fontSize="md"
+                borderRadius={"full"}
+                minWidth={"200px"}
+                color={"white"}
+                marginTop={"15px"}
+                marginBottom={"10px"}
+                padding={"20px"}
+                // semi transparent white outline
+                outline={"1px solid rgba(255, 255, 255, 0.6)"}
+                style={{
+                  background:
+                    "linear-gradient(45deg, #007BFF, #3F60D9, #5E43BA, #7C26A5, #9A0A90)",
+                  backgroundSize: "300% 300%",
+                  animation: "Gradient 10s infinite linear",
+                }}
+              >
+                Upload
+                <style jsx>{`
+                  @keyframes Gradient {
+                    0% {
+                      background-position: 100% 0%;
+                    }
+                    50% {
+                      background-position: 0% 100%;
+                    }
+                    100% {
+                      background-position: 100% 0%;
+                    }
+                  }
+                `}</style>
+              </Button>
+            </VStack>
+          </form>
+        </VStack>
+      </Box>
+    </>
+  );
 };
 
 export default CreateEpisode;
